Remove "installed" flag and add automatic detection
If you keep KEm in a folder on usb flash or desktop, you near-always
expect that everything will live in this folder. If your KEm is somehow
ended in C:\Program Files\, you probably don't want it to be portable,
and windows won't allow you that anyway.

So i deleted this flag and added checks against some well-knows paths.