Minor deobfuscation of mmf-related classes.
This was done primarily to integrate the freej2me-plus implementation of the MMF player in the future. The original one is proprietary and compiled for Win32. FJ2ME is not accurate at all, but it's better than nothing. So for Win32, the original stays, while the multiplatform version will use the new one.